+++ .gitignore
@@ -7,8 +7,8 @@ demo/*
 *.json
 
 # Demo files
-/user_tool/process-supervisor/
-/user_tool/process-supervisor/* 
+process-supervisor/
+process-supervisor/* 
 
 user_tool/__pycache__
 user_tool/__pycache__/*
@@ -17,3 +17,56 @@ user_tool/.coverage
 user_tool/.coverage*
 user_tool/htmlcov
 user_tool/htmlcov/*
+
+# Byte-compiled / optimized / DLL files
+__pycache__/
+*.py[cod]
+*$py.class
+
+# C extensions
+*.so
+
+# Distribution / packaging
+.Python
+build/
+develop-eggs/
+dist/
+downloads/
+eggs/
+.eggs/
+lib/
+lib64/
+parts/
+sdist/
+var/
+wheels/
+share/python-wheels/
+*.egg-info/
+.installed.cfg
+*.egg
+MANIFEST
+
+# PyInstaller
+#  Usually these files are written by a python script from a template
+#  before PyInstaller builds the exe, so as to inject date/other infos into it.
+*.manifest
+*.spec
+
+# Installer logs
+pip-log.txt
+pip-delete-this-directory.txt
+
+# Unit test / coverage reports
+htmlcov/
+.tox/
+.nox/
+.coverage
+.coverage.*
+.cache
+nosetests.xml
+coverage.xml
+*.cover
+*.py,cover
+.hypothesis/
+.pytest_cache/
+cover/
